chore: general improvements and refactor of headless-ssr-commerce
https://coveord.atlassian.net/browse/KIT-3681
