+++ packages/samples/headless-ssr-commerce/app/layout.tsx
@@ -13,7 +13,6 @@ export default function RootLayout({children}: {children: React.ReactNode}) {
         <div style={{display: 'flex', alignItems: 'center', gap: '10px'}}>
           <Link href={'/search'}>Search Page</Link>
           <Link href={'/listing'}>Listing Page</Link>
-          <Link href={'/recommendation'}>Recommendations</Link>
         </div>
 
         {children}
