+++ packages/samples/headless-ssr-commerce/components/recent-queries.tsx
@@ -1,7 +1,4 @@
-import {
-  useInstantProducts,
-  useRecentQueriesList,
-} from '../_lib/commerce-engine';
+import {useInstantProducts, useRecentQueriesList} from '../lib/commerce-engine';
 
 export default function RecentQueries() {
   const {state, controller} = useRecentQueriesList();
