+++ packages/samples/headless-ssr-commerce/components/standalone-search-box.tsx
@@ -6,7 +6,7 @@ import {
   useInstantProducts,
   useRecentQueriesList,
   useStandaloneSearchBox,
-} from '../_lib/commerce-engine';
+} from '../lib/commerce-engine';
 import InstantProducts from './instant-product';
 import RecentQueries from './recent-queries';
