+++ packages/samples/headless-ssr-commerce/components/triggers/triggers.tsx
