+++ packages/samples/headless-ssr-commerce/lib/navigatorContextProvider.ts
